Frdyy

+++ info111/arbres-hauteur.cpp
@@ -15,8 +15,8 @@ using namespace std;
     // nom = 15
 
 int main() {
-    vector<vector<string>> t = litTableauCSV("donnees/arbresremarquablesparis.csv", 35  );
-    double hauteur_max = 0.0;
+    vector<vector<string>> t = litTableauCSV("donnees/arbresremarquablesparis.csv", 35);
+    int hauteur_max = 0;
     int index_max = 0;
     vector<string> C = colonne(t, 8);
     vector<int> tableau_hauteur = conversionInt(C);
